cleanup bokeh module, split dependencies and utils
issues: closes # 2041
