+++ bokeh/util/serialization.py
@@ -1,24 +1,19 @@
+"""
+
+"""
 from __future__ import absolute_import
 
 import logging
-logger = logging.getLogger(__name__)
-
-from functools import reduce
-import math
-import platform
-import sys
-import uuid
-import flask
-import json
-
-from six.moves.urllib.parse import urljoin as sys_urljoin
-from six import iteritems
+log = logging.getLogger(__name__)
 
 _simple_id = 1000
 
 def make_id():
     global _simple_id
-    from . import settings
+
+    import uuid
+    from ..settings import settings
+
     if settings.simple_ids(False):
         _simple_id += 1
         new_id = _simple_id
@@ -27,52 +22,33 @@ def make_id():
     return str(new_id)
 
 def urljoin(*args):
+    from six.moves.urllib.parse import urljoin as sys_urljoin
+    from functools import reduce
     return reduce(sys_urljoin, args)
 
 def get_json(response):
-    """unifying retrieving json from an http response,for requests <1.0, >1.0, and
+    """unifying retrieving json from an http response, for requests <1.0, >1.0, and
     flask test client
     """
+    import json
+    import flask
     if isinstance(response, flask.Response):
-        #flask testing
+        # flask testing
         return json.loads(response.data.decode('utf-8'))
     else:
-        #requests
+        # requests
         if hasattr(response.json, '__call__'):
             return response.json()
         else:
             return response.json
 
-def encode_utf8(u):
-    if sys.version_info[0] == 2:
-        u = u.encode('utf-8')
-    return u
-
-def decode_utf8(u):
-    if sys.version_info[0] == 2:
-        u = u.decode('utf-8')
-    return u
-
-_scales = [1e0, 1e3, 1e6, 1e9]
-_units = ['s', 'ms', 'us', 'ns']
-
-def scale_delta(time):
-    if time > 0.0:
-        order = min(-int(math.floor(math.log10(time)) // 3), 3)
-    else:
-        order = 3
-
-    return time*_scales[order], _units[order]
-
-def is_py3():
-    return sys.version_info[0] == 3
-
-def is_pypy():
-    return platform.python_implementation() == "PyPy"
 
 def convert_references(json_obj):
+
+    import iteritems
     from .plot_object import PlotObject
     from .properties import HasProps
+
     def convert(obj):
         if isinstance(obj, PlotObject):
             return obj.ref
@@ -80,6 +56,7 @@ def convert_references(json_obj):
             return obj.to_dict()
         else:
             return obj
+
     def helper(json_obj):
         if isinstance(json_obj, list):
             for idx, x in enumerate(json_obj):
@@ -87,7 +64,9 @@ def convert_references(json_obj):
         if isinstance(json_obj, dict):
             for k, x in iteritems(json_obj):
                 json_obj[k] = convert(x)
+
     json_apply(json_obj, helper)
+
     return json_obj
 
 def dump(objs, docid, changed_only=True):
@@ -104,29 +83,6 @@ def dump(objs, docid, changed_only=True):
         ref["attributes"].update({"id": ref["id"], "doc" : docid})
         json_objs.append(ref)
     return json_objs
-
-def nice_join(seq, sep=", "):
-    seq = [str(x) for x in seq]
-
-    if len(seq) <= 1:
-        return sep.join(seq)
-    else:
-        return "%s or %s" % (sep.join(seq[:-1]), seq[-1])
-
-def publish_display_data(data, source='bokeh'):
-    """Compatibility wrapper for IPython publish_display_data which removes the
-    `source` (first) argument in later versions.
-
-    Parameters
-    ----------
-    source : str
-    data : dict
-    """
-    import IPython.core.displaypub as displaypub
-    try:
-        displaypub.publish_display_data(source, data)
-    except TypeError:
-        displaypub.publish_display_data(data)
 
 def is_ref(frag):
     return isinstance(frag, dict) and \
@@ -159,6 +115,6 @@ def resolve_json(fragment, models):
         if fragment['id'] in models:
             return models[fragment['id']]
         else:
-            logging.error("model not found for %s", fragment)
+            log.error("model not found for %s", fragment)
             return None
     return json_apply(fragment, check_func, func)
